+++ src/CoreEx.Cosmos/Batch/CosmosDbBatch.cs
@@ -67,7 +67,7 @@ namespace CoreEx.Cosmos.Batch
         /// <param name="cancellationToken">The <see cref="CancellationToken"/>.</param>
         /// <remarks>Each item is added individually and is not transactional.</remarks>
         public static Task ImportBatchAsync<T, TModel>(this CosmosDbContainer<T, TModel> container, IEnumerable<TModel> items, Func<TModel, TModel>? modelUpdater = null, CosmosDbArgs? dbArgs = null, CancellationToken cancellationToken = default) where T : class, IEntityKey, new() where TModel : class, IEntityKey, new()
-            => ImportBatchAsync(container?.CosmosDb!, container?.Container.Id!, items, modelUpdater, dbArgs ?? container.ThrowIfNull().DbArgs, cancellationToken);
+            => ImportBatchAsync(container.ThrowIfNull(nameof(container)).CosmosDb!, container.Container.Id!, items, modelUpdater, dbArgs ?? container.DbArgs, cancellationToken);
 
         /// <summary>
         /// Imports (creates) a batch of named items from the <paramref name="jsonDataReader"/> into the specified <paramref name="containerId"/>.
@@ -106,7 +106,7 @@ namespace CoreEx.Cosmos.Batch
         /// <returns><c>true</c> indicates that one or more items were deserialized and imported; otherwise, <c>false</c> for none found.</returns>
         /// <remarks>Each item is added individually and is not transactional.</remarks>
         public static Task<bool> ImportBatchAsync<T, TModel>(this CosmosDbContainer<T, TModel> container, JsonDataReader jsonDataReader, string? name = null, Func<TModel, TModel>? modelUpdater = null, CosmosDbArgs? dbArgs = null, CancellationToken cancellationToken = default) where T : class, IEntityKey, new() where TModel : class, IEntityKey, new()
-            => ImportBatchAsync(container?.CosmosDb!, container?.Container.Id!, jsonDataReader, name, modelUpdater, dbArgs ?? container.ThrowIfNull().DbArgs, cancellationToken);
+            => ImportBatchAsync(container.ThrowIfNull(nameof(container)).CosmosDb!, container.Container.Id!, jsonDataReader, name, modelUpdater, dbArgs ?? container.DbArgs, cancellationToken);
 
         /// <summary>
         /// Imports (creates) a batch of <see cref="CosmosDbValue{TModel}"/> <paramref name="items"/>.
@@ -131,7 +131,7 @@ namespace CoreEx.Cosmos.Batch
             foreach (var item in items)
             {
                 var cdv = new CosmosDbValue<TModel>(item);
-                ((ICosmosDbValue)cdv).PrepareBefore(dbArgs.Value);
+                ((ICosmosDbValue)cdv).PrepareBefore(dbArgs.Value, null);
 
                 if (SequentialExecution)
                     await container.CreateItemAsync(modelUpdater?.Invoke(cdv) ?? cdv, dbArgs.Value.PartitionKey, dbArgs.Value.ItemRequestOptions, cancellationToken).ConfigureAwait(false);
@@ -154,7 +154,15 @@ namespace CoreEx.Cosmos.Batch
         /// <param name="cancellationToken">The <see cref="CancellationToken"/>.</param>
         /// <remarks>Each item is added individually and is not transactional.</remarks>
         public static Task ImportValueBatchAsync<T, TModel>(this CosmosDbValueContainer<T, TModel> container, IEnumerable<TModel> items, Func<CosmosDbValue<TModel>, CosmosDbValue<TModel>>? modelUpdater = null, CosmosDbArgs? dbArgs = null, CancellationToken cancellationToken = default) where T : class, IEntityKey, new() where TModel : class, IEntityKey, new()
-            => ImportValueBatchAsync(container?.CosmosDb!, container?.Container.Id!, items, modelUpdater, dbArgs ?? container.ThrowIfNull().DbArgs, cancellationToken);
+        {
+            CosmosDbValue<TModel> func(CosmosDbValue<TModel> cvm)
+            {
+                cvm.Type = container.ModelContainer.TypeName;
+                return modelUpdater?.Invoke(cvm) ?? cvm;
+            }
+
+            return ImportValueBatchAsync(container.ThrowIfNull(nameof(container)).CosmosDb!, container.Container.Id!, items, func, dbArgs ?? container.DbArgs, cancellationToken);
+        }
 
         /// <summary>
         /// Imports (creates) a batch of named <see cref="CosmosDbValue{TModel}"/> items from the <paramref name="jsonDataReader"/> into the specified <paramref name="containerId"/>.
@@ -193,7 +201,15 @@ namespace CoreEx.Cosmos.Batch
         /// <returns><c>true</c> indicates that one or more items were deserialized and imported; otherwise, <c>false</c> for none found.</returns>
         /// <remarks>Each item is added individually and is not transactional.</remarks>
         public static Task<bool> ImportValueBatchAsync<T, TModel>(this CosmosDbValueContainer<T, TModel> container, JsonDataReader jsonDataReader, string? name = null, Func<CosmosDbValue<TModel>, CosmosDbValue<TModel>>? modelUpdater = null, CosmosDbArgs? dbArgs = null, CancellationToken cancellationToken = default) where T : class, IEntityKey, new() where TModel : class, IEntityKey, new()
-            => ImportValueBatchAsync(container?.CosmosDb!, container?.Container.Id!, jsonDataReader, name, modelUpdater, dbArgs ?? container.ThrowIfNull().DbArgs, cancellationToken);
+        {
+            CosmosDbValue<TModel> func(CosmosDbValue<TModel> cvm)
+            {
+                cvm.Type = container.ModelContainer.TypeName;
+                return modelUpdater?.Invoke(cvm) ?? cvm;
+            }
+
+            return ImportValueBatchAsync(container.ThrowIfNull(nameof(container)).CosmosDb!, container.Container.Id!, jsonDataReader, name ?? container.ModelContainer.TypeName, (Func<CosmosDbValue<TModel>, CosmosDbValue<TModel>>)func, dbArgs ?? container.DbArgs, cancellationToken);
+        }
 
         /// <summary>
         /// Imports (creates) a batch of named <see cref="CosmosDbValue{TModel}"/> items from the <paramref name="jsonDataReader"/> into the specified <paramref name="containerId"/>.
@@ -224,7 +240,7 @@ namespace CoreEx.Cosmos.Batch
                     foreach (var item in items.Where(x => x is not null))
                     {
                         var cdv = Activator.CreateInstance(t, item)!;
-                        ((ICosmosDbValue)cdv).PrepareBefore(dbArgs.Value);
+                        ((ICosmosDbValue)cdv).PrepareBefore(dbArgs.Value, null);
 
                         if (SequentialExecution)
                             await container.CreateItemAsync(modelUpdater?.Invoke(cdv) ?? cdv, dbArgs.Value.PartitionKey, dbArgs.Value.ItemRequestOptions, cancellationToken).ConfigureAwait(false);